move sigusr1-related test to a posix-only file
